Fix print heap formatting
When this function was written, \n created a new line and went to the start of it. Now, this has been split into both \n and \r, so this function was causing the new lines to continue off in the same column as the last one. Adding the \r's fixes this.