Free linked lists; fix heap bugs
I've suspected the heap was broken for a while, but it was rarely used so it didn't end up causing any bugs. However, now with the heap being used a ton for the new types of strings (every character in a Char_List is now a new object on the heap) the bugs all became apparent. Now, however, it should all work.