Change malloc/free to use offsets/smaller max allocation size
This should save some heap memory from the header, useful for things like linked lists where a lot of the memory usage comes just from those headers. Might reduce the size of some of those types even more, since not using nearly that much, but for now this seems like a safe first step. In the future, if more is needed, changing should not be too difficult since just changing the typedef should be enough.